# dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserSignUpResponse": {
				"properties": {
					"email": {
						"type": "string",
						"description": "가입한 회원 이메일"
					},
					"name": {
						"type": "string",
						"description": "가입한 회원 이름"
					},
					"preferCategory": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "선호 음식 카테고리 이름 목록"
					}
				},
				"required": [
					"email",
					"name",
					"preferCategory"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiSuccessResponse_UserSignUpResponse_": {
				"properties": {
					"isSuccess": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false,
						"description": "요청 성공 여부"
					},
					"code": {
						"type": "string",
						"description": "응답 코드"
					},
					"message": {
						"type": "string",
						"description": "응답 메시지"
					},
					"result": {
						"$ref": "#/components/schemas/UserSignUpResponse",
						"description": "응답 데이터"
					}
				},
				"required": [
					"isSuccess",
					"code",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiErrorResponse": {
				"properties": {
					"isSuccess": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false,
						"description": "요청 성공 여부"
					},
					"code": {
						"type": "string",
						"description": "에러 코드"
					},
					"message": {
						"type": "string",
						"description": "에러 메시지"
					},
					"result": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true,
						"description": "에러 상세 데이터"
					}
				},
				"required": [
					"isSuccess",
					"code",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserSignUpRequest": {
				"properties": {
					"email": {
						"type": "string",
						"description": "로그인에 사용할 이메일"
					},
					"password": {
						"type": "string",
						"description": "로그인에 사용할 비밀번호"
					},
					"name": {
						"type": "string",
						"description": "회원 이름"
					},
					"gender": {
						"type": "string",
						"description": "회원 성별"
					},
					"birth": {
						"type": "string",
						"description": "생년월일"
					},
					"address": {
						"type": "string",
						"description": "주소"
					},
					"detailAddress": {
						"type": "string",
						"description": "상세 주소"
					},
					"phoneNumber": {
						"type": "string",
						"description": "연락처"
					},
					"preferences": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array",
						"description": "선호 음식 카테고리 ID 목록"
					}
				},
				"required": [
					"email",
					"password",
					"name",
					"gender",
					"birth",
					"phoneNumber",
					"preferences"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TokenResponse": {
				"properties": {
					"accessToken": {
						"type": "string",
						"description": "API 요청에 사용할 Access Token"
					},
					"refreshToken": {
						"type": "string",
						"description": "Access Token 재발급에 사용할 Refresh Token"
					}
				},
				"required": [
					"accessToken",
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiSuccessResponse_TokenResponse_": {
				"properties": {
					"isSuccess": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false,
						"description": "요청 성공 여부"
					},
					"code": {
						"type": "string",
						"description": "응답 코드"
					},
					"message": {
						"type": "string",
						"description": "응답 메시지"
					},
					"result": {
						"$ref": "#/components/schemas/TokenResponse",
						"description": "응답 데이터"
					}
				},
				"required": [
					"isSuccess",
					"code",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserLoginRequest": {
				"properties": {
					"email": {
						"type": "string",
						"description": "로그인에 사용할 이메일"
					},
					"password": {
						"type": "string",
						"description": "로그인에 사용할 비밀번호"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RefreshTokenRequest": {
				"properties": {
					"refreshToken": {
						"type": "string",
						"description": "로그인 시 발급받은 Refresh Token"
					}
				},
				"required": [
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserProfileResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"gender": {
						"type": "string"
					},
					"birth": {
						"type": "string",
						"format": "date-time"
					},
					"address": {
						"type": "string"
					},
					"detailAddress": {
						"type": "string",
						"nullable": true
					},
					"phoneNumber": {
						"type": "string"
					},
					"preferCategory": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"email",
					"name",
					"gender",
					"birth",
					"address",
					"detailAddress",
					"phoneNumber",
					"preferCategory"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiSuccessResponse_UserProfileResponse_": {
				"properties": {
					"isSuccess": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false,
						"description": "요청 성공 여부"
					},
					"code": {
						"type": "string",
						"description": "응답 코드"
					},
					"message": {
						"type": "string",
						"description": "응답 메시지"
					},
					"result": {
						"$ref": "#/components/schemas/UserProfileResponse",
						"description": "응답 데이터"
					}
				},
				"required": [
					"isSuccess",
					"code",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserProfileUpdateRequest": {
				"properties": {
					"name": {
						"type": "string",
						"description": "회원 이름"
					},
					"gender": {
						"type": "string",
						"description": "회원 성별"
					},
					"birth": {
						"type": "string",
						"description": "생년월일"
					},
					"address": {
						"type": "string",
						"description": "주소"
					},
					"detailAddress": {
						"type": "string",
						"description": "상세 주소"
					},
					"phoneNumber": {
						"type": "string",
						"description": "연락처"
					},
					"preferences": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array",
						"description": "선호 음식 카테고리 ID 목록"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ApiSuccessResponse_null_": {
				"properties": {
					"isSuccess": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false,
						"description": "요청 성공 여부"
					},
					"code": {
						"type": "string",
						"description": "응답 코드"
					},
					"message": {
						"type": "string",
						"description": "응답 메시지"
					},
					"result": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true,
						"description": "응답 데이터"
					}
				},
				"required": [
					"isSuccess",
					"code",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AddReviewRequest": {
				"properties": {
					"rating": {
						"type": "number",
						"format": "double",
						"description": "리뷰 평점 (0 이상 5 이하)"
					},
					"content": {
						"type": "string",
						"description": "리뷰 내용"
					}
				},
				"required": [
					"rating",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AddMissionRequest": {
				"properties": {
					"title": {
						"type": "string",
						"description": "미션 제목"
					},
					"description": {
						"type": "string",
						"description": "미션 설명"
					},
					"rewardPoint": {
						"type": "number",
						"format": "double",
						"description": "미션 보상 포인트"
					},
					"deadline": {
						"type": "string",
						"description": "미션 마감일"
					}
				},
				"required": [
					"title",
					"description",
					"rewardPoint",
					"deadline"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StoreSummary": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "가게 ID"
					},
					"name": {
						"type": "string",
						"description": "가게 이름"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "리뷰 ID"
					},
					"userId": {
						"type": "number",
						"format": "double",
						"description": "리뷰 작성 회원 ID"
					},
					"storeId": {
						"type": "number",
						"format": "double",
						"description": "리뷰 대상 가게 ID"
					},
					"rating": {
						"type": "number",
						"format": "double",
						"description": "리뷰 평점"
					},
					"content": {
						"type": "string",
						"description": "리뷰 내용"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"description": "리뷰 작성 시각"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"description": "리뷰 수정 시각"
					},
					"store": {
						"$ref": "#/components/schemas/StoreSummary",
						"description": "리뷰 대상 가게"
					}
				},
				"required": [
					"id",
					"userId",
					"storeId",
					"rating",
					"content",
					"createdAt",
					"updatedAt",
					"store"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewListResponse": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/ReviewResponse"
						},
						"type": "array",
						"description": "리뷰 목록"
					},
					"pagination": {
						"properties": {
							"cursor": {
								"type": "number",
								"format": "double",
								"nullable": true
							}
						},
						"required": [
							"cursor"
						],
						"type": "object",
						"description": "다음 조회에 사용할 커서"
					}
				},
				"required": [
					"data",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiSuccessResponse_ReviewListResponse_": {
				"properties": {
					"isSuccess": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false,
						"description": "요청 성공 여부"
					},
					"code": {
						"type": "string",
						"description": "응답 코드"
					},
					"message": {
						"type": "string",
						"description": "응답 메시지"
					},
					"result": {
						"$ref": "#/components/schemas/ReviewListResponse",
						"description": "응답 데이터"
					}
				},
				"required": [
					"isSuccess",
					"code",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StoreMissionResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "미션 ID"
					},
					"storeId": {
						"type": "number",
						"format": "double",
						"description": "가게 ID"
					},
					"title": {
						"type": "string",
						"description": "미션 제목"
					},
					"description": {
						"type": "string",
						"description": "미션 설명"
					},
					"rewardPoint": {
						"type": "number",
						"format": "double",
						"description": "미션 보상 포인트"
					},
					"deadline": {
						"type": "string",
						"format": "date-time",
						"description": "미션 마감일"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"description": "생성 시각"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"description": "수정 시각"
					}
				},
				"required": [
					"id",
					"storeId",
					"title",
					"description",
					"rewardPoint",
					"deadline",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiSuccessResponse_StoreMissionResponse-Array_": {
				"properties": {
					"isSuccess": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false,
						"description": "요청 성공 여부"
					},
					"code": {
						"type": "string",
						"description": "응답 코드"
					},
					"message": {
						"type": "string",
						"description": "응답 메시지"
					},
					"result": {
						"items": {
							"$ref": "#/components/schemas/StoreMissionResponse"
						},
						"type": "array",
						"description": "응답 데이터"
					}
				},
				"required": [
					"isSuccess",
					"code",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserMissionResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "회원 미션 ID"
					},
					"userId": {
						"type": "number",
						"format": "double",
						"description": "회원 ID"
					},
					"missionId": {
						"type": "number",
						"format": "double",
						"description": "미션 ID"
					},
					"status": {
						"type": "string",
						"description": "회원 미션 상태"
					},
					"completedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true,
						"description": "미션 완료 시각"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"description": "생성 시각"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"description": "수정 시각"
					}
				},
				"required": [
					"id",
					"userId",
					"missionId",
					"status",
					"completedAt",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiSuccessResponse_UserMissionResponse-or-null_": {
				"properties": {
					"isSuccess": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false,
						"description": "요청 성공 여부"
					},
					"code": {
						"type": "string",
						"description": "응답 코드"
					},
					"message": {
						"type": "string",
						"description": "응답 메시지"
					},
					"result": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserMissionResponse"
							}
						],
						"nullable": true,
						"description": "응답 데이터"
					}
				},
				"required": [
					"isSuccess",
					"code",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "week8",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users/signup": {
			"post": {
				"operationId": "SignUp",
				"responses": {
					"201": {
						"description": "회원가입 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessResponse_UserSignUpResponse_"
								}
							}
						}
					},
					"400": {
						"description": "회원가입 필수값 누락",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"description": "회원가입 API",
				"summary": "신규 회원을 생성하거나 이미 존재하는 회원 정보를 갱신합니다.",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserSignUpRequest"
							}
						}
					}
				}
			}
		},
		"/users/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessResponse_TokenResponse_"
								}
							}
						}
					},
					"400": {
						"description": "로그인 필수값 누락",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "로그인 실패",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"description": "로그인 API",
				"summary": "이메일과 비밀번호로 로그인하고 JWT Access/Refresh Token을 발급합니다.",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserLoginRequest"
							}
						}
					}
				}
			}
		},
		"/users/refresh": {
			"post": {
				"operationId": "Refresh",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessResponse_TokenResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Refresh Token 누락",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "유효하지 않은 토큰",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"description": "Access Token 재발급 API",
				"summary": "Refresh Token으로 새로운 Access/Refresh Token을 발급합니다.",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RefreshTokenRequest"
							}
						}
					}
				}
			}
		},
		"/users/me": {
			"get": {
				"operationId": "GetMe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessResponse_UserProfileResponse_"
								}
							}
						}
					},
					"401": {
						"description": "로그인 필요",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"description": "내 정보 조회 API",
				"summary": "Bearer Token으로 로그인한 사용자의 정보를 조회합니다.",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateMe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessResponse_UserProfileResponse_"
								}
							}
						}
					},
					"401": {
						"description": "로그인 필요",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"description": "내 정보 수정 API",
				"summary": "Bearer Token으로 로그인한 사용자의 전화번호, 생일, 선호 카테고리 등을 수정합니다.",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserProfileUpdateRequest"
							}
						}
					}
				}
			}
		},
		"/stores/{storeId}/reviews": {
			"post": {
				"operationId": "AddReview",
				"responses": {
					"201": {
						"description": "리뷰 등록 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessResponse_null_"
								}
							}
						}
					},
					"400": {
						"description": "가게 ID 또는 리뷰 필수값 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "로그인 필요",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 가게",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"description": "리뷰 등록 API",
				"summary": "Bearer Token으로 로그인한 사용자가 특정 가게에 리뷰를 등록합니다.",
				"tags": [
					"Store"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddReviewRequest"
							}
						}
					}
				}
			}
		},
		"/stores/{storeId}/missions": {
			"post": {
				"operationId": "AddMission",
				"responses": {
					"201": {
						"description": "미션 등록 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessResponse_null_"
								}
							}
						}
					},
					"400": {
						"description": "가게 ID 또는 미션 필수값 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "로그인 필요",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 가게",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"description": "미션 등록 API",
				"summary": "Bearer Token으로 로그인한 사용자만 특정 가게에 미션을 등록할 수 있습니다.",
				"tags": [
					"Store"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddMissionRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetStoreMissions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessResponse_StoreMissionResponse-Array_"
								}
							}
						}
					},
					"400": {
						"description": "가게 ID 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"description": "가게 미션 목록 조회 API",
				"summary": "특정 가게의 미션 목록을 조회합니다.",
				"tags": [
					"Store"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/reviews/my": {
			"get": {
				"operationId": "GetMyReviews",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessResponse_ReviewListResponse_"
								}
							}
						}
					},
					"400": {
						"description": "커서 값 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "로그인 필요",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"description": "내 리뷰 목록 조회 API",
				"summary": "Bearer Token으로 로그인한 사용자가 작성한 리뷰 목록을 커서 기반으로 조회합니다.",
				"tags": [
					"Store"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/missions/{userMissionId}/complete": {
			"patch": {
				"operationId": "CompleteMission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessResponse_UserMissionResponse-or-null_"
								}
							}
						}
					},
					"400": {
						"description": "회원 미션 ID 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "로그인 필요",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "진행 중인 미션 없음",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"409": {
						"description": "이미 완료한 미션",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"description": "미션 완료 API",
				"summary": "Bearer Token으로 로그인한 사용자의 진행 중인 미션을 완료 상태로 변경합니다.",
				"tags": [
					"Store"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userMissionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/missions/{missionId}": {
			"patch": {
				"operationId": "Challenge",
				"responses": {
					"201": {
						"description": "미션 도전 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessResponse_null_"
								}
							}
						}
					},
					"400": {
						"description": "미션 ID 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "로그인 필요",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 미션",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"409": {
						"description": "이미 진행 중이거나 완료한 미션",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"description": "미션 도전 API",
				"summary": "Bearer Token으로 로그인한 사용자가 특정 미션에 도전합니다.",
				"tags": [
					"Mission"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "missionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}